Rename submodule basic_strategy to blackjack to avoid name shadowing.
