+++ src/blackjack/__main__.py
@@ -5,9 +5,9 @@ import datetime
 import pandas as pd
 
 from src import TABLE_DTYPES, engine
-from src.basic_strategy.card_eval import hand_eval
-from src.basic_strategy.hand import Hand
-from src.basic_strategy.mode_selector import deal_solo_cards
+from src.blackjack.card_eval import hand_eval
+from src.blackjack.hand import Hand
+from src.blackjack.mode_selector import deal_solo_cards
 
 ENDC = "\033[0m"
 OKGREEN = "\033[92m"
